refactor: separate out functions to make firmware loader main file more manageable
